fix bugs

+++ Test/TestProgram-3.cpp
@@ -5,7 +5,7 @@ int add(int a, int b){
     return a + b;
 }
 
-uint64_t x = 1, y, z = add(1,2);
+uint32_t x = 1, y, z = add(1,2);
 
 int main(){
     y = x + 21;
